+++ api-v2/src/main/kotlin/com/cordtables/v2/repositories/AdminRepositories.kt
@@ -1,0 +1,8 @@
+package com.cordtables.v2.repositories
+
+import com.cordtables.v2.models.AdminPeople
+import org.springframework.data.repository.CrudRepository
+import org.springframework.stereotype.Repository
+
+@Repository
+interface AdminPeopleRepository : CrudRepository<AdminPeople, String>
