+++ api-v2/src/main/resources/sql/common.schema.v1.sql
@@ -1,0 +1,38 @@
+insert into admin_table_name_enum(value) values
+  ('common_blogs'),
+  ('common_blog_posts'),
+  ('common_cell_channels'),
+  ('common_coalition_memberships'),
+  ('common_coalitions'),
+  ('common_directories'),
+  ('common_discussion_channels'),
+  ('common_education_by_person'),
+  ('common_education_entries'),
+  ('common_file_versions'),
+  ('common_files'),
+  ('common_languages'),
+  ('common_locations'),
+  ('common_notes'),
+  ('common_organizations'),
+  ('common_org_chart_positions'),
+  ('common_org_chart_position_graph'),
+  ('common_people_graph'),
+  ('common_people_to_org_relationships'),
+  ('common_posts'),
+  ('common_up.prayer_requests'),
+  ('common_up.prayer_notifications'),
+  ('common_scripture_references'),
+  ('common_site_text_strings'),
+  ('common_site_text_translations'),
+  ('common_stage_graph'),
+  ('common_stage_notifications'),
+  ('common_stage_role_column_grants'),
+  ('common_stages'),
+  ('common_threads'),
+  ('common_ticket_assignments'),
+  ('common_ticket_feedback'),
+  ('common_ticket_graph'),
+  ('common_tickets'),
+  ('common_work_estimates'),
+  ('common_work_records'),
+  ('common_workflows');
