+++ api-v2/src/main/resources/sql/sc.schema.v1.sql
@@ -1,0 +1,44 @@
+
+insert into admin_table_name_enum(value) values
+  ('sc_budget_records'),
+  ('sc_budget_records_partnerships'),
+  ('sc_budgets'),
+  ('sc_ceremonies'),
+  ('sc_change_to_plans'),
+  ('sc_ethno_arts'),
+  ('sc_ethnologue'),
+  ('sc_field_regions'),
+  ('sc_field_zones'),
+  ('sc_file_versions'),
+  ('sc_films'),
+  ('sc_funding_accounts'),
+  ('sc_global_partner_assessments'),
+  ('sc_global_partner_engagements'),
+  ('sc_global_partner_engagement_people'),
+  ('sc_global_partner_performance'),
+  ('sc_global_partner_transitions'),
+  ('sc_internship_engagements'),
+  ('sc_known_languages_by_person'),
+  ('sc_language_engagements'),
+  ('sc_language_goal_definitions'),
+  ('sc_language_goals'),
+  ('sc_language_locations'),
+  ('sc_languages'),
+  ('sc_locations'),
+  ('sc_organization_locations'),
+  ('sc_organizations'),
+  ('sc_partners'),
+  ('sc_partnerships'),
+  ('sc_people'),
+  ('sc_periodic_reports'),
+  ('sc_periodic_reports_directory'),
+  ('sc_person_unavailabilities'),
+  ('sc_pinned_projects'),
+  ('sc_posts'),
+  ('sc_posts_directory'),
+  ('sc_product_scripture_references'),
+  ('sc_products'),
+  ('sc_project_locations'),
+  ('sc_project_members'),
+  ('sc_projects'),
+  ('sc_stories');
