+++ api-v2/src/main/resources/sql/sil.schema.v1.sql
@@ -1,0 +1,11 @@
+insert into admin_table_name_enum(value) values
+  ('sil_country_codes'),
+  ('sil_language_codes'),
+  ('sil_language_index'),
+  ('sil_iso_639_3'),
+  ('sil_iso_639_3_names'),
+  ('sil_iso_639_3_macrolanguages'),
+  ('sil_iso_639_3_retirements'),
+  ('sil_table_of_countries'),
+  ('sil_table_of_languages'),
+  ('sil_table_of_languages_in_country');
